HBX-1587: Move class 'org.hibernate.cfg.JDBCCollectionSecondPass' to 'org.hibernate.tool.internal.reveng.JdbcCollectionSecondPass'
