+++ main/src/main/java/org/hibernate/tool/internal/reveng/JdbcCollectionSecondPass.java
@@ -1,9 +1,11 @@
-package org.hibernate.cfg;
+package org.hibernate.tool.internal.reveng;
 
+import java.lang.reflect.Field;
 import java.util.Map;
 
 import org.hibernate.MappingException;
 import org.hibernate.boot.spi.MetadataBuildingContext;
+import org.hibernate.cfg.CollectionSecondPass;
 import org.hibernate.mapping.Collection;
 import org.hibernate.mapping.DependantValue;
 import org.hibernate.mapping.OneToMany;
@@ -11,23 +13,23 @@ import org.hibernate.mapping.PersistentClass;
 import org.hibernate.mapping.Value;
 
 @SuppressWarnings("serial")
-public class JDBCCollectionSecondPass extends CollectionSecondPass {
+public class JdbcCollectionSecondPass extends CollectionSecondPass {
 
 	MetadataBuildingContext mdbc;
 
-    JDBCCollectionSecondPass(MetadataBuildingContext mdbc, Collection coll) {
+    public JdbcCollectionSecondPass(MetadataBuildingContext mdbc, Collection coll) {
         super(mdbc, coll);
         this.mdbc = mdbc;
     }
 
    @SuppressWarnings("rawtypes")
    public void secondPass(Map persistentClasses, Map inheritedMetas) throws MappingException {
-        bindCollectionSecondPass(collection, persistentClasses, mdbc, inheritedMetas);
+        bindCollectionSecondPass(getCollection(), persistentClasses, mdbc, inheritedMetas);
     }
 
     @SuppressWarnings("rawtypes")
 	public void doSecondPass(Map persistentClasses) throws MappingException {
-    	Value element = collection.getElement();
+    	Value element = getCollection().getElement();
     	DependantValue elementDependantValue = null;
     	String oldElementForeignKeyName = null;
     	if(element instanceof DependantValue) {
@@ -35,7 +37,7 @@ public class JDBCCollectionSecondPass extends CollectionSecondPass {
 			oldElementForeignKeyName = elementDependantValue.getForeignKeyName();
     		elementDependantValue.setForeignKeyName("none"); // Workaround to avoid DependantValue to create foreignkey just because reference columns are not the same + no need to create keys already in the db!
     	}
-    	Value key = collection.getKey();
+    	Value key = getCollection().getKey();
     	DependantValue keyDependantValue = null;
     	String oldKeyForeignKeyName = null;
     	if (key instanceof DependantValue) {
@@ -67,6 +69,20 @@ public class JDBCCollectionSecondPass extends CollectionSecondPass {
 
             oneToMany.setAssociatedClass(persistentClass); // Child
         }
+    }
+    
+    private Collection getCollection() {
+    	try {
+    		Field field = getClass().getSuperclass().getDeclaredField("collection");
+    		field.setAccessible(true);
+    		return (Collection)field.get(this);
+    	} catch (NoSuchFieldException e) {
+    		// this will happen if the implementation of the superclass changes
+    		throw new RuntimeException(e);
+    	} catch (IllegalAccessException e) {
+    		// this should not happen
+    		throw new RuntimeException(e);
+    	}
     }
     
 }
